backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-multipart
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
